# api/docs/swagger.yaml
basePath: /
definitions:
...
  processes.Info:
    properties:
      description:
        type: string
      id:
        type: string
      jobControlOptions:
        items:
          type: string
        type: array
      outputTransmission:
        items:
          type: string
        type: array
      title:
        type: string
      version:
        type: string
    type: object
  processes.Input:
    properties:
      literalDataDomain:
        $ref: '#/definitions/processes.LiteralDataDomain'
    type: object
  processes.Inputs:
    properties:
      description:
        type: string
      id:
        type: string
      input:
        $ref: '#/definitions/processes.Input'
      maxOccurs:
        type: integer
      minOccurs:
        type: integer
      title:
        type: string
    type: object
  processes.Link:
    properties:
      href:
        type: string
      rel:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  processes.LiteralDataDomain:
    properties:
      dataType:
        type: string
      valueDefinition:
        $ref: '#/definitions/processes.ValueDefinition'
    type: object
  processes.Output:
    properties:
      formats:
        items:
          type: string
        type: array
    type: object
  processes.Outputs:
    properties:
      description:
        type: string
      id:
        type: string
      inputID:
        description: json omit
        type: string
      output:
        $ref: '#/definitions/processes.Output'
      title:
        type: string
    type: object
  processes.Resources:
    properties:
      cpus:
        type: number
      memory:
        type: integer
    type: object
  processes.ValueDefinition:
    properties:
      anyValue:
        type: boolean
      possibleValues:
        items:
          type: string
        type: array
    type: object
  processes.processDescription:
    properties:
      image:
        type: string
      info:
        $ref: '#/definitions/processes.Info'
      inputs:
        items:
          $ref: '#/definitions/processes.Inputs'
        type: array
      links:
        items:
          $ref: '#/definitions/processes.Link'
        type: array
      maxResources:
        $ref: '#/definitions/processes.Resources'
      outputs:
        items:
          $ref: '#/definitions/processes.Outputs'
        type: array
    type: object
...
host: localhost:5050
info:
  contact:
    email: slawler@dewberry.com
    name: Seth Lawler
  description: An OGC compliant process server.
  title: SEPEX Server
  version: dev-8.16.23
paths:
...
  /processes/{processID}:
    get:
      consumes:
      - '*/*'
      description: '[Process Description Specification](https://docs.ogc.org/is/18-062r2/18-062r2.html#sc_process_description)'
      parameters:
      - description: 'example: pyecho'
        in: path
        name: processID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/processes.processDescription'
      summary: Describe Process Information
      tags:
      - processes
...
schemes:
- http
swagger: "2.0"